--- components/auth/UserAccountControl.module.scss
@@ -1,8 +1,0 @@
-.buttonSpread {
-  button {
-    margin: 0.25em;
-  }
-  button:nth-child(1) {
-    margin-left: 0;
-  }
-}
