+++ lib/data/useAllGroups.ts
@@ -3,9 +3,8 @@ import { DocumentReference } from '@cougargrades/types/dist/FirestoreStubs'
 import { CourseInstructorResult } from './useCourseData'
 import { getBadges } from './getBadges'
 import { defaultComparator, descendingComparator } from '../../components/datatable'
-import { firebaseApp, getFirestoreDocument } from '../ssg'
 
-type AllGroupsResultItem = { [key: string]: GroupResult[] };
+export type AllGroupsResultItem = { [key: string]: GroupResult[] };
 
 export interface AllGroupsResult {
   categories: string[];
@@ -83,56 +82,6 @@ export function course2Result(data: CoursePlus): CourseInstructorResult {
 
 export const ALL_GROUPS_SENTINEL = 'All Groups'
 
-export async function getAllGroups(): Promise<AllGroupsResult> {
-  const db = firebaseApp.firestore()
-  const query = db.collection('groups').where('categories', 'array-contains', '#UHCoreCurriculum')
-  const querySnap = await query.get()
-  const data: Group[] = querySnap.docs.filter(e => e.exists).map(e => e.data() as Group);
-  
-  // sanitize output
-  for(let i = 0; i < data.length; i++) {
-    data[i].courses = []
-    data[i].sections = []
-  }
-
-  // make categories
-  const categories = [
-    ...(
-      Array.from(new Set(data.map(e => Array.isArray(e.categories) ? e.categories.filter(cat => !cat.startsWith('#')) : []).flat()))
-        .sort((a,b) => defaultComparator(a,b)) // [ '(All)', '(2022-2023)', '(2021-2022)', '(2020-2021)' ]
-        .slice(0,2) // don't endlessly list the groups, they're still accessible from a course directly
-      ),
-    //ALL_GROUPS_SENTINEL
-  ];
-
-  // make a key/value store of category -> GroupResult[]
-  const results = categories
-    .reduce((obj, key) => {
-      if(key === ALL_GROUPS_SENTINEL) {
-        // obj[key] = [
-        //   ...(status === 'success' ? data.filter(e => Array.isArray(e.categories) && e.categories.length === 0).map(e => group2Result(e)) : [])
-        // ];
-      }
-      else {
-        obj[key] = [
-          ...(data.filter(e => Array.isArray(e.categories) && e.categories.includes(key)).map(e => group2Result(e)))
-        ];
-      }
-      return obj;
-    }, {} as AllGroupsResultItem);
-
-  return {
-    categories,
-    results,
-    core_curriculum: [
-      ...(data.filter(e => Array.isArray(e.categories) && e.categories.includes('UH Core Curriculum')).map(e => group2Result(e)))
-    ],
-    all_groups: [
-      ...(data.filter(e => Array.isArray(e.categories) && ! e.categories.includes('UH Core Curriculum')).map(e => group2Result(e)))
-    ],
-  }
-}
-
 export interface GroupPlus extends Group {
   courseCount?: number;
   sectionCount?: number;
@@ -143,50 +92,4 @@ export interface CoursePlus extends Course {
   instructorCount?: number;
 }
 
-export async function getOneGroup(groupId: string, includeSections: boolean = false): Promise<PopulatedGroupResult | undefined> {
-  const data = await getFirestoreDocument<GroupPlus>(`/groups/${groupId}`)
-  const didLoadCorrectly = data !== undefined && typeof data === 'object' && Object.keys(data).length > 1
 
-  const settledData = await Promise.allSettled([
-    (data && Array.isArray(data.courses) && Util.isDocumentReferenceArray(data.courses) ? Util.populate<CoursePlus>(data.courses) : Promise.resolve<CoursePlus[]>([])),
-    (data && includeSections && Array.isArray(data?.sections) && Util.isDocumentReferenceArray(data.sections) ? Util.populate<Section>(data.sections, 10, true) : Promise.resolve<Section[]>([])),
-  ]);
-  const [courseDataSettled, sectionDataSettled] = settledData
-  const courseData = courseDataSettled.status === 'fulfilled' ? courseDataSettled.value : [];
-  const sectionData = sectionDataSettled.status === 'fulfilled' ? sectionDataSettled.value : [];
-
-  if (didLoadCorrectly) {
-    data.keywords = []
-    data.courseCount = Array.isArray(data.courses) ? data.courses.length : 0
-    data.courses = courseData
-      // filter out undefined because there might be some empty references
-      .filter(e => e !== undefined)
-      // sort courses by total enrolled
-      .sort((a,b) => b.enrollment.totalEnrolled - a.enrollment.totalEnrolled)
-      // sanitize unwanted document references
-      .map(course => ({ 
-        ...course,
-        // property necessary for some client-side calculations
-        sectionCount: Array.isArray(course.sections) ? course.sections.length : 0,
-        sections: [],
-        instructorCount: Array.isArray(course.instructors) ? course.instructors.length : 0,
-        instructors: [],
-        // property not needed
-        //sections: Array.isArray(course.sections) ? course.sections.map(sec => ({ id: sec?.id as any as string })) as any : [],
-        //instructors: Array.isArray(course.instructors) ? course.instructors.map(ins => ({ id: ins?.id as any as string })) as any : [],
-        groups: [],
-        keywords: [],
-      }));
-    data.sectionCount = Array.isArray(data.sections) ? data.sections.length : 0
-    data.sections = sectionData
-      // filter out undefined because there might be some empty references
-      .filter(e => e !== undefined)
-      // sanitize unwanted document references
-      .map(sec => ({ 
-        ...sec,
-        instructors: [],
-      }));
-  }
-
-  return didLoadCorrectly ? group2PopResult(data) : undefined
-}
