+++ lib/hook.tsx
@@ -1,6 +1,5 @@
-import { useEffect, useRef } from 'react'
+import React from 'react'
 import { useMedia } from 'react-use'
-import { useFirebaseApp } from 'reactfire'
 
 export function useIsMobile() {
   return useMedia('(max-width: 576px)');
@@ -8,15 +7,4 @@ export function useIsMobile() {
 
 export function useIsCondensed() {
   return useMedia('(max-width: 768px)');
-}
-
-export function useAnalyticsRef() {
-  const firebaseApp = useFirebaseApp()
-  const analyticsRef = useRef<firebase.default.analytics.Analytics>(null)
-
-  useEffect(() => {
-    analyticsRef.current = firebaseApp.analytics()
-  },[])
-
-  return analyticsRef;
 }
