+++ package.json
@@ -21,7 +21,8 @@
     "@mui/material": "^5.2.6",
     "@primer/css": "^19.1.1",
     "bootstrap": "^5.0.1",
-    "firebase": "^8.10.0",
+    "firebase": "^9.16.0",
+    "firebase-admin": "^9.12.0",
     "fitty": "^2.3.5",
     "gray-matter": "^4.0.3",
     "next": "~12.0.3",
@@ -39,7 +40,7 @@
     "react-tilty": "^2.0.3",
     "react-transition-group": "^4.4.2",
     "react-use": "^17.2.4",
-    "reactfire": "^3.0.0",
+    "reactfire": "^4.2.2",
     "recoil": "^0.5.2",
     "rehype-raw": "^6.1.1",
     "rehype-sanitize": "^5.0.1",
