+++ pages/api/course/[courseName].ts
@@ -1,7 +1,8 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { CACHE_CONTROL } from '../../../lib/cache'
-import { CourseResult, getCourseData } from '../../../lib/data/useCourseData'
+import { getCourseData } from '../../../lib/data/back/getCourseData';
+import { CourseResult } from '../../../lib/data/useCourseData'
 import { extract } from '../../../lib/util'
 
 export default async function GetCourseData(req: NextApiRequest, res: NextApiResponse<CourseResult>) {
