+++ pages/api/group/[groupId]/sections.ts
@@ -1,7 +1,8 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { CACHE_CONTROL } from '../../../../lib/cache'
-import { PopulatedGroupResult, getOneGroup } from '../../../../lib/data/useAllGroups'
+import { getOneGroup } from '../../../../lib/data/back/getOneGroup';
+import { PopulatedGroupResult } from '../../../../lib/data/useAllGroups'
 import { extract } from '../../../../lib/util'
 
 export default async function GetOneGroup(req: NextApiRequest, res: NextApiResponse<PopulatedGroupResult>) {
