+++ pages/api/instructor/[instructorName].ts
@@ -1,7 +1,8 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { CACHE_CONTROL } from '../../../lib/cache'
-import { InstructorResult, getInstructorData } from '../../../lib/data/useInstructorData'
+import { getInstructorData } from '../../../lib/data/back/getInstructorData'
+import { InstructorResult } from '../../../lib/data/useInstructorData'
 import { extract } from '../../../lib/util'
 
 export default async function GetInstructorData(req: NextApiRequest, res: NextApiResponse<InstructorResult>) {
